+++ HISTORY.md
@@ -5,6 +5,9 @@
 ### Public API changes
 * Replaced ColumnFamilyOptions::table_properties_collectors with ColumnFamilyOptions::table_properties_collector_factories
 
+### New Features
+* Hash index for block-based table will be materialized and reconstructed more efficiently. Previously hash index is constructed by scanning the whole table during every table open.
+
 ## 3.0.0 (05/05/2014)
 
 ### Public API changes